Show ERROR on evaluation errors
An evaluation error would occur if division by zero is attempted, which
will either result to NaN or Infinity (or -Infinity). The app allowed
chaining operators to these values, but throw an error afterwards
because letters are not recognized.

The inputEquals function now catches these errors then prints them to
the console. It will then return null for the input state. From here on
null represents an evaluation error, and will be displayed as such by
the calculator.