Add keydown listeners for calculator buttons
I renamed onClick to onActivate, because the button can now be activated
the same with clicking and pressing a keyboard key.

Now I'm contemplating making helper components for the buttons (like
CalculatorButton.Digit, or CalculatorButton.Operator). I did that before
but I scrapped it because it seemed overkill. But now, not so much.